Simplify by adding pooling classes
