+++ ehr2vec/evaluation/encodings.py
@@ -3,14 +3,16 @@ from os.path import join, split
 import numpy as np
 import pandas as pd
 import torch
+from common.logger import TqdmToLogger
+from common.config import instantiate
 from dataloader.collate_fn import dynamic_padding
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 from trainer.trainer import EHRTrainer
-from common.logger import TqdmToLogger
+
 
 class Forwarder(EHRTrainer):
-    def __init__(self, model, dataset, batch_size=50, stop_iter=None, pool_method="mean", logger=None, test=False, layers='last', run=None):
+    def __init__(self, model, dataset, batch_size=50, stop_iter=None, pool_method="mean", pooler=None, logger=None, test=False, layers='last', run=None):
         self.model = model
         self.logger = logger
         self.model.eval()
@@ -24,7 +26,7 @@ class Forwarder(EHRTrainer):
         self.validate_parameters()
         self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
         self.dataloader = DataLoader(self.dataset, batch_size=self.batch_size, shuffle=False, collate_fn=dynamic_padding)
-        
+        self.pooler =  instantiate(pooler)
     def forward_patients(self)->dict:
         encodings = []
         with torch.no_grad():
@@ -34,23 +36,13 @@ class Forwarder(EHRTrainer):
                 self.to_device(batch)
                 output = self.forward_pass(batch)
                 hidden = output.last_hidden_state.detach()
-                pooled_vec = self.pool(hidden, mask, output)
+                pooled_vec = self.pooler.pool(hidden, mask, output)
                 encodings.append(pooled_vec)
                 if self.test:
                     if i>1:
                         break
         return torch.cat(encodings)
     
-    def pool(self, hidden, mask, output):
-        if self.pool_method=='mean':
-            return torch.mean(mask.unsqueeze(-1) * hidden, dim=1)
-        elif self.pool_method=='weighted_sum':
-            return self.attention_weighted_sum(output, hidden, mask, self.layers)
-        elif self.pool_method=='CLS':
-            return hidden[:,0,:] # CLS token
-        else:
-            raise ValueError(f"Method {self.pool_method} not implemented yet.")
-        
     
     def validate_parameters(self):
         valid_methods = ['mean', 'weighted_sum', 'CLS']
@@ -59,21 +51,6 @@ class Forwarder(EHRTrainer):
             raise ValueError(f"Method {self.pool_method} not implemented yet.")
         if self.layers not in valid_layers:
             raise ValueError(f"Layers {self.layers} not implemented yet.")
-
-    @staticmethod
-    def attention_weighted_sum(outputs, hidden, mask, layers='all'):
-        """Compute embedding using attention weights"""
-        attention = outputs['attentions'] # tuple num layers (batch_size, num_heads, sequence_length, sequence_length)
-        if layers=='all':
-            attention = torch.stack(attention).mean(dim=0).mean(dim=1) # average over all layers and heads
-        elif layers=='last':
-            attention = attention[-1].mean(dim=1) # average over all layers and heads
-        else:
-            raise ValueError(f"Layers {layers} not implemented yet.")
-        weights = torch.mean(attention, dim=1) # (batch_size, sequence_length, sequence_length)
-        weights = weights / torch.sum(weights, dim=1, keepdim=True) # normalize, potentially uise softmax
-        hidden_vec = torch.sum(mask.unsqueeze(-1) * hidden * weights.unsqueeze(-1), dim=1)
-        return hidden_vec
 
     def produce_patient_trajectory_embeddings(self, start_code_id=4)->dict:
         """Produce embedding for each patient. Showing an increasing number of codes to the model,
